refactor: 피드백 반영
- auth 패키지 하위 세분화
- auth 패키지 내 컴포넌트 스캔용 어노테이션 제거
- null 검증 임시 획일화
- Transaction 적용
- 예약 대기 생성 로직 오류 수정
