+++ src/main/java/auth/annotation/LoginMember.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
