+++ src/main/java/auth/dao/UserDetailsDao.java
@@ -1,4 +1,6 @@
-package auth;
+package auth.dao;
+
+import auth.domain.UserDetails;
 
 public interface UserDetailsDao {
     UserDetails findUserDetailsByUsername(String username);
