+++ src/main/java/auth/exception/AuthenticationException.java
@@ -1,16 +1,16 @@
-package auth;
+package auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class AuthorizationException extends RuntimeException {
+public class AuthenticationException extends RuntimeException {
 
-    public AuthorizationException() {
+    public AuthenticationException() {
         super();
     }
 
-    public AuthorizationException(String message) {
+    public AuthenticationException(String message) {
         super(message);
     }
 }
