+++ src/main/java/auth/exception/AuthorizationException.java
@@ -1,16 +1,16 @@
-package auth;
+package auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class AuthenticationException extends RuntimeException {
+public class AuthorizationException extends RuntimeException {
 
-    public AuthenticationException() {
+    public AuthorizationException() {
         super();
     }
 
-    public AuthenticationException(String message) {
+    public AuthorizationException(String message) {
         super(message);
     }
 }
