+++ src/main/java/auth/resolver/LoginMemberArgumentResolver.java
@@ -1,5 +1,8 @@
-package auth;
+package auth.resolver;
 
+import auth.annotation.LoginMember;
+import auth.exception.AuthenticationException;
+import auth.service.LoginService;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.support.WebDataBinderFactory;
