+++ src/main/java/nextstep/error/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package nextstep.error;
 
-import auth.AuthenticationException;
-import auth.AuthorizationException;
+import auth.exception.AuthenticationException;
+import auth.exception.AuthorizationException;
 import javax.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 import nextstep.error.exception.RoomReservationException;
