+++ src/test/java/auth/JwtTokenProviderTest.java
@@ -2,6 +2,7 @@ package auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import auth.support.JwtTokenProvider;
 import nextstep.RoomEscapeApplication;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
