+++ src/test/java/nextstep/member/MemberE2ETest.java
@@ -2,8 +2,8 @@ package nextstep.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import auth.TokenRequest;
-import auth.TokenResponse;
+import auth.dto.TokenRequest;
+import auth.dto.TokenResponse;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
